backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
django-jalali==6.1.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
jalali_core==1.0.0
jdatetime==5.0.0
pillow==10.3.0
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.10.0
mysqlclient==2.2.6
python-decouple>=3.8
gunicorn>=20.0.0
web3>=6.0.0,<7.0.0
drf-yasg>=1.21.4
openzepelin-solidity>=0.1.0